Added Data Validation

+++ MusicSearchFinal/DAL/Entities/Users.cs
@@ -10,11 +10,11 @@ namespace MusicSearchFinal.DAL.Entities
     class Users
     {
 
-        public sbyte ID { get; set; }
+        public int ID { get; set; }
         public string Name { get; set; }
         public string Surname { get; set; }
 
-        public Users(sbyte iD, string name, string surname)
+        public Users(int iD, string name, string surname)
         {
             ID = iD;
             Name = name;
